backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend application by using Django"
authors = ["tnakagami"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "^4.2.11"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
uvicorn = "^0.29.0"
python-magic = "^0.4.27"
beautifulsoup4 = "^4.12.3"
langchain = "^0.1.16"
langgraph = "^0.0.38"
langchain-openai = "^0.1.3"
langchain-anthropic = "^0.1.11"
langchain-aws = "^0.1.1"
langchain-fireworks = "^0.1.2"
langchain-google-vertexai = "^1.0.1"
langchain-robocorp = "^0.0.5"
tomli-w = "^1.0.0"
boto3 = "^1.34.88"
duckduckgo-search = "^5.3.0"
arxiv = "^2.1.0"
kay = "^0.1.2"
wikipedia = "^1.4.0"
fireworks-ai = "^0.14.0"
xmltodict = "^0.13.0"
unstructured = { extras = ["doc", "docx"], version = "^0.13.2", optional = true }
pdfminer-six = "^20231228"
pgvector = "^0.2.5"
psycopg = "^3.1.18"
django-asgi-lifespan = "^0.3.1"
django-view-breadcrumbs = "^2.5.1"
django-bootstrap-icons = "^0.8.7"
django-markdownx = "^4.0.7"
django-picklefield = "^3.2.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-clarity = "^1.0.1"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
pytest-env = "^1.1.3"
pytest-factoryboy = "^2.7.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.14.0"
pytest-randomly = "^3.15.0"
factory-boy = "^3.3.0"
mock = "^5.1.0"
coverage = "^7.4.4"
django-webtest = "^1.9.11"

[tool.pytest_env]
PYTHONPATH = "/opt/app"

[tool.pytest.ini_options]
addopts = "-vv --rootdir=/opt/app --cov=. --cov-report=xml --cov-report=html --cov-config=.coveragerc"
DJANGO_SETTINGS_MODULE = "config.settings.development"
testpaths = ["all_tests"]
python_files = ["test_*.py", "tests.py", "*_test.py"]
markers = [
  "account: marks tests as account app",
  "chatbot: marks tests as chatbot app",
  "model: marks tests as Model",
  "form: marks tests as Form",
  "view: marks tests as View",
  "drf: marks tests as Django REST framework",
  "private: marks tests as Private class/method/function",
  "util: marks tests as Utility class/method/function",
  "langgraph: marks tests as LangGraph related classes",
]
cache_dir = "/opt/home/.cache"
filterwarnings = [
  "ignore::DeprecationWarning",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
